【シーンマネージャー】プレイシーンを作成しました。

+++ 10Jam/Children.cpp
@@ -17,5 +17,5 @@ void Children::Update(XMFLOAT2& PlayerPos)
 
 void Children::Draw()
 {
-	DrawCircle(pos.x, pos.y, 8.0f, GetColor(255, 255, 255), true);
+	DrawCircle((int)pos.x, (int)pos.y, (int)8.0f, GetColor(255, 255, 255), true);
 }
